Keep unsupported datatype so that it is possible to process in python, based on review comment
